refactor: move os schemas serializers da views para serializers
